Secure views in store and profile app by adding login required decorators

+++ store/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -69,8 +70,13 @@ def product_detail(request, product_id):
     return render(request, 'store/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store via the frontend """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only store owners can add product to site')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -89,8 +95,13 @@ def add_product(request):
     return render(request, 'store/add_product.html', context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only store owners can edit a product')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -107,8 +118,13 @@ def edit_product(request, product_id):
     return render(request, 'store/edit_product.html', context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ delete a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only store owners can delete a product')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'product deleted!')
